backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
PyPDF2
transformers
torch
flask-mail
sentencepiece
 